# api-contracts/openapi/platform-sdk-management-api-v1.yaml
openapi: 3.0.3
info:
  title: Keiko Platform-SDK Management API
  description: |
    **Architektur-konforme Management API für Platform-SDK Kommunikation**
    
    Diese API ermöglicht SDK-Management und -Konfiguration durch:
    - Agent-Registrierung und -Verwaltung
    - Configuration Management
    - Health Monitoring und Status-Updates
    - Function Call Orchestration
    
    **Compliance:** Strikte API-basierte Kommunikation ohne direkte Backend-Dependencies.
  version: 1.0.0
  contact:
    name: Keiko Architecture Team
    email: architecture@keiko.dev
servers:
  - url: https://platform-api.keiko.dev/v1
    description: Production Platform API
  - url: https://staging-platform-api.keiko.dev/v1
    description: Staging Platform API
  - url: http://localhost:8000/v1
    description: Local Development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
  /agents:
    get:
      summary: List Agents
      description: Abrufen aller registrierten Agents
      operationId: listAgents
      tags:
        - Agent Management
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: ["active", "inactive", "error", "maintenance"]
        - name: agent_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'

    post:
      summary: Register Agent
      description: |
        SDK registriert neuen Agent bei Platform.
        Ersetzt direkte Backend-Registrierung durch API-Call.
      operationId: registerAgent
      tags:
        - Agent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
      responses:
        '201':
          description: Agent successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          description: Invalid agent data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AgentRegistrationRequest:
      type: object
      required:
        - agent_id
        - agent_type
        - capabilities
      properties:
        agent_id:
          type: string
          pattern: '^agent_[a-zA-Z0-9_]+$'
          description: Eindeutige Agent-ID
        agent_type:
          type: string
          enum:
            - "code_assistant"
            - "data_analyst"
            - "workflow_manager"
            - "monitoring_agent"
        capabilities:
          type: array
          items:
            type: string
          description: Agent-Fähigkeiten
        metadata:
          type: object
          additionalProperties:
            type: string
        config:
          type: object
          description: Initial agent configuration
...
    AgentResponse:
      type: object
      required:
        - agent_id
        - agent_type
        - status
        - created_at
      properties:
        agent_id:
          type: string
        agent_type:
          type: string
        status:
          type: string
          enum: ["active", "inactive", "error", "maintenance"]
        capabilities:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
    AgentListResponse:
      type: object
      required:
        - agents
        - total
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...